requirements.txt
python-telegram-bot[job-queue]==21.7
requests==2.32.3
qrcode[pil]==7.4.2
python-dotenv==1.0.1
psutil==5.9.8